refactor: Break up functions into smaller files
